Error Circular Buffers: Now tasks have warnings and errors being sent to USB Controller (#71)

+++ Middlewares/CircularBuffer/Inc/CircularBufferWriter.hpp
@@ -10,41 +10,41 @@ template <typename T>
 class CircularBufferWriter
 {
 public:
-    CircularBufferWriter(T* buffer, uint32_t* writerIndex, uint32_t size);
+    CircularBufferWriter(T* buffer, size_t* writerIndex, size_t size);
 
     // Index management
-    uint32_t GetIndex() const;
-    void SetIndex(uint32_t index);
+    size_t GetIndex() const;
+    void SetIndex(size_t index);
 
     // Element write
     void WriteElement(const T& value);
-    void WriteElement(uint32_t index, const T& value);
+    void WriteElement(size_t index, const T& value);
     void WriteElementAndIncrementIndex(const T& value);
 
 private:
     T* _buffer;        // external buffer memory
-    uint32_t* _writerIndex;
-    uint32_t _size;
+    size_t* _writerIndex;
+    size_t _size;
 };
 
 template <typename T>
-inline CircularBufferWriter<T>::CircularBufferWriter(T* buffer, uint32_t* writerIndex, uint32_t size)
+inline CircularBufferWriter<T>::CircularBufferWriter(T* buffer, size_t* writerIndex, size_t size)
     : _buffer(buffer), _writerIndex(writerIndex), _size(size)
 {
 	*_writerIndex = 0;
 }
 
 template <typename T>
-inline uint32_t CircularBufferWriter<T>::GetIndex() const
+inline size_t CircularBufferWriter<T>::GetIndex() const
 {
     taskENTER_CRITICAL(); 
-    uint32_t writerIndex = *_writerIndex;
+    size_t writerIndex = *_writerIndex;
     taskEXIT_CRITICAL(); 
     return writerIndex;
 }
 
 template <typename T>
-inline void CircularBufferWriter<T>::SetIndex(uint32_t index)
+inline void CircularBufferWriter<T>::SetIndex(size_t index)
 {
     taskENTER_CRITICAL(); 
     *_writerIndex = index % _size;
@@ -61,7 +61,7 @@ inline void CircularBufferWriter<T>::WriteElement(const T& value)
 }
 
 template <typename T>
-inline void CircularBufferWriter<T>::WriteElement(uint32_t index, const T& value)
+inline void CircularBufferWriter<T>::WriteElement(size_t index, const T& value)
 {
     taskENTER_CRITICAL(); 
     _buffer[index % _size] = value;
